DEPI_Project_App/requirements.txt
# List your dependencies here
streamlit==1.28.0
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
requests==2.31.0
matplotlib==3.7.2

# Plotly compatible with Kaleido
plotly==5.15.0
kaleido==0.2.1

ultralytics==8.1.0
torch==2.1.0

# For environment variables (.env)
python-dotenv